🧱 New Communication class
